requirements.txt
astroid==2.0.4
attrs==17.4.0
autopep8==1.4
boto3==1.9.23
botocore==1.12.23
certifi==2018.8.24
chardet==3.0.4
colorama==0.3.9
coverage==4.4.2
dj-database-url==0.4.2
Django==2.1.5
django-bootstrap4==0.0.5
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-storages==1.7.1
django-markdownx
docutils==0.14
gunicorn==19.7.1
idna==2.7
isort==4.3.4
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.3.0
pipenv==2018.10.9
pluggy==0.6.0
py==1.5.2
pycodestyle==2.4.0
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
pytest==3.3.2
pytest-cov==2.5.1
pytest-django==3.1.2
python-dateutil==2.7.3
python-decouple==3.1
python-resize-image==1.1.18
pytz==2017.3
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
sqlparse==0.2.4
typed-ast==1.1.0
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
whitenoise==3.3.1
wrapt==1.10.11
